Refactor codebase to follow style guide
