+++ src/index.html
@@ -1,22 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Minimal Tower Defense</title>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-  <body>
-    <div id="hud">
-      <span id="lives">Lives: 10</span>
-      <span id="gold">Gold: 20</span>
-      <span id="wave">Wave: 1/5</span>
-      <span id="status"></span>
-      <button id="nextWave">Next Wave</button>
-      <button id="placeTower">Place Tower (10)</button>
-      <span id="status"></span>
-      <button id="restart">Restart</button>
-    </div>
-    <canvas id="game" width="800" height="450"></canvas>
-    <script src="game.js"></script>
-  </body>
+    <head>
+        <meta charset="UTF-8" />
+        <title>Minimal Tower Defense</title>
+        <link rel="stylesheet" href="style.css" />
+    </head>
+    <body>
+        <div id="hud">
+            <span id="lives">Lives: 10</span>
+            <span id="gold">Gold: 20</span>
+            <span id="wave">Wave: 1/5</span>
+            <span id="status"></span>
+            <button id="nextWave">Next Wave</button>
+            <button id="placeTower">Place Tower (10)</button>
+            <button id="restart">Restart</button>
+        </div>
+        <canvas id="game" width="800" height="450"></canvas>
+        <script type="module" src="main.js"></script>
+    </body>
 </html>
