+++ src/main.js
@@ -1,0 +1,5 @@
+import Game from './Game.js';
+
+const canvas = document.getElementById('game');
+const game = new Game(canvas);
+game.run();
