+++ src/style.css
@@ -1,16 +1,16 @@
 body {
-  margin: 0;
-  background: #fff;
+    margin: 0;
+    background: #fff;
 }
 
 #hud {
-  display: flex;
-  gap: 10px;
-  align-items: center;
-  padding: 10px;
-  font-family: sans-serif;
+    display: flex;
+    gap: 10px;
+    align-items: center;
+    padding: 10px;
+    font-family: sans-serif;
 }
 
 #placeTower.active {
-  background: #ff0;
+    background: #ff0;
 }
